+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Added
 
+-   Add middle-click to close tabs. (#1351)
 -   Add vim emulation with [custom commands](https://cpeditor.org/docs/preferences/code-edit/#custom-vim-commands). (#220 and #1270)
 
 ### Changed